+++ contracts/access/Ownable.sol
@@ -4,6 +4,8 @@
 pragma solidity ^0.8.20;
 
 import {Context} from "../utils/Context.sol";
+import {IERC173} from "../interfaces/IERC173.sol";
+import {IERC165, ERC165} from "../utils/introspection/ERC165.sol";
 
 /**
  * @dev Contract module which provides a basic access control mechanism, where
@@ -17,30 +19,14 @@ import {Context} from "../utils/Context.sol";
  * `onlyOwner`, which can be applied to your functions to restrict their use to
  * the owner.
  */
-abstract contract Ownable is Context {
+abstract contract Ownable is Context, IERC173, ERC165 {
     address private _owner;
 
-    /**
-     * @dev The caller account is not authorized to perform an operation.
-     */
+    /// @dev The caller account is not authorized to perform an operation.
     error OwnableUnauthorizedAccount(address account);
 
-    /**
-     * @dev The owner is not a valid owner account. (eg. `address(0)`)
-     */
+    /// @dev The owner is not a valid owner account. (eg. `address(0)`)
     error OwnableInvalidOwner(address owner);
-
-    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
-
-    /**
-     * @dev Initializes the contract setting the address provided by the deployer as the initial owner.
-     */
-    constructor(address initialOwner) {
-        if (initialOwner == address(0)) {
-            revert OwnableInvalidOwner(address(0));
-        }
-        _transferOwnership(initialOwner);
-    }
 
     /**
      * @dev Throws if called by any account other than the owner.
@@ -50,42 +36,39 @@ abstract contract Ownable is Context {
         _;
     }
 
-    /**
-     * @dev Returns the address of the current owner.
-     */
+    /// @dev Initializes the contract setting the address provided by the deployer as the initial owner.
+    constructor(address initialOwner) {
+        if (initialOwner == address(0)) {
+            revert OwnableInvalidOwner(address(0));
+        }
+        _transferOwnership(initialOwner);
+    }
+
+    /// @inheritdoc ERC165
+    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC165, IERC165) returns (bool) {
+        return interfaceId == type(IERC173).interfaceId || super.supportsInterface(interfaceId);
+    }
+
+    /// @inheritdoc IERC173
     function owner() public view virtual returns (address) {
         return _owner;
     }
 
     /**
-     * @dev Throws if the sender is not the owner.
-     */
-    function _checkOwner() internal view virtual {
-        if (owner() != _msgSender()) {
-            revert OwnableUnauthorizedAccount(_msgSender());
-        }
-    }
-
-    /**
-     * @dev Leaves the contract without owner. It will not be possible to call
-     * `onlyOwner` functions. Can only be called by the current owner.
+     * @dev See {IERC173-transferOwnership}.
      *
      * NOTE: Renouncing ownership will leave the contract without an owner,
      * thereby disabling any functionality that is only available to the owner.
      */
-    function renounceOwnership() public virtual onlyOwner {
-        _transferOwnership(address(0));
+    function transferOwnership(address newOwner) public virtual override onlyOwner {
+        _transferOwnership(newOwner);
     }
 
-    /**
-     * @dev Transfers ownership of the contract to a new account (`newOwner`).
-     * Can only be called by the current owner.
-     */
-    function transferOwnership(address newOwner) public virtual onlyOwner {
-        if (newOwner == address(0)) {
-            revert OwnableInvalidOwner(address(0));
+    /// @dev Throws if the sender is not the owner.
+    function _checkOwner() internal view virtual {
+        if (owner() != _msgSender()) {
+            revert OwnableUnauthorizedAccount(_msgSender());
         }
-        _transferOwnership(newOwner);
     }
 
     /**